+++ Intersectii/point2D.h
@@ -4,18 +4,18 @@
 class Point2D {
 
 public:
-	Point2D(const float & x = 0.0, const float & y = 0.0) :
+	Point2D(const double & x = 0.0, const double & y = 0.0) :
 			X(x), Y(y) {
 	}
-	float X;
-	float Y;
+	double X;
+	double Y;
 	bool operator==(const Point2D & other) const;
 	bool operator<(const Point2D & other) const;
 	bool operator>(const Point2D & other) const;
 	bool operator<=(const Point2D & other) const;
 	bool operator>=(const Point2D & other) const;
 	bool operator!=(const Point2D & other) const;
-	float Distance(const Point2D & other) const;
+	double Distance(const Point2D & other) const;
 
 };
